Folder restructure for front end. Add new entry point for webpack to account for footer component. Create header module.
