+++ react-client/src/footer/app.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import Header from './components/Header.jsx';
+import Footer from './components/Footer.jsx';
 
 class App extends React.Component {
   constructor(props) {
@@ -10,16 +9,13 @@ class App extends React.Component {
     }
   }
 
-  componentDidMount() {
-  }
-
   render () {
     return (
     <div>
-      <Header />
+      <Footer />
     </div>
     )
   }
 }
 
-ReactDOM.render(<App />, document.getElementById('header'));
+export default App;
