+++ react-client/src/footer/components/Footer.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from '../styles/footer.module.css';
+
+const Footer = () => (
+  <div className={styles.footer}>
+    <div>this is a footer!</div>
+  </div>
+)
+
+export default Footer;
