+++ react-client/src/footer/styles/footer.module.css
@@ -1,0 +1,3 @@
+.footer {
+  margin-top: 90%;
+}
