+++ react-client/src/header/index.jsx
@@ -1,0 +1,5 @@
+import ReactDOM from 'react-dom';
+import React from 'react';
+import App from './app.jsx';
+
+ReactDOM.render(<App />, document.getElementById('header'));
