+++ react-client/src/header/styles/header.module.css
@@ -1,8 +1,8 @@
 #head {
   width: 100%;
   margin: 0 auto;
-  top:0;
-  left:0;
+  top: 0;
+  left: 0;
   height: 5.5%;
   position: absolute;
   background-color: rgba(245, 245, 245, 0.952);
@@ -108,12 +108,14 @@
 }
 
 .nikelogo {
+  cursor: pointer;
   margin-left: 3%;
   margin-bottom: 2%;
 }
 
 .men {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -123,6 +125,7 @@
 
 .men:hover {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -134,6 +137,7 @@
 
 .newRelease {
   float: right;
+  cursor: pointer;
   width: fit-content;
   position: relative;
   font-size: 16.5px;
@@ -144,6 +148,7 @@
 
 .newRelease:hover {
   float: right;
+  cursor: pointer;
   width: fit-content;
   position: relative;
   font-size: 16.5px;
@@ -156,6 +161,7 @@
 
 .women {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -165,6 +171,7 @@
 
 .women:hover {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -176,6 +183,7 @@
 
 .kids {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -185,6 +193,7 @@
 
 .kids:hover {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -196,6 +205,7 @@
 
 .customize {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -205,6 +215,7 @@
 
 .customize:hover {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -216,6 +227,7 @@
 
 .sale {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
@@ -225,6 +237,7 @@
 
 .sale:hover {
   float: right;
+  cursor: pointer;
   position: relative;
   font-size: 16.5px;
   font-family: Helvetica;
