+++ react-client/src/header/styles/headerBottom.module.css
@@ -1,0 +1,36 @@
+#container {
+  width: 100%;
+  margin: 0 auto;
+  top: 15%;
+  left: 0;
+  height: 10%;
+  position: absolute;
+  background-color: rgba(245, 245, 245, 0.952);
+  color: black;
+  transition: top 0.3s;
+}
+
+.text {
+  align-items: center;
+  text-align: center;
+  position: relative;
+  margin-top: 1.6%;
+  font-size: large;
+  font-family: Helvetica;
+}
+
+.rightArrow {
+  cursor: pointer;
+  position: relative;
+  float: right;
+  margin-right: 20%;
+  margin-top: -1.3%;
+}
+
+.leftArrow {
+  cursor: pointer;
+  position: relative;
+  float: left;
+  margin-left: 20%;
+  margin-top: -1.3%;
+}
