+++ webpack.config.js
@@ -5,14 +5,16 @@
 /* eslint-disable key-spacing */
 /* eslint-disable no-var */
 /* eslint-disable import/newline-after-import */
-var path = require('path');
-var SRC_DIR = path.join(__dirname, '/react-client/src');
-var DIST_DIR = path.join(__dirname, '/react-client/dist');
+const path = require('path');
+const SRC_DIR = path.join(__dirname, '/react-client/src');
+const DIST_DIR = path.join(__dirname, '/react-client/dist');
 
 module.exports = {
-  entry: `${SRC_DIR}/index.jsx`,
+  entry: {
+    bundle: [`${SRC_DIR}/header/index.jsx`, `${SRC_DIR}/footer/index.jsx`],
+  },
   output: {
-    filename: 'bundle.js',
+    filename: '[name].js',
     path: DIST_DIR
   },
   module : {
